backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
websockets
python-dotenv